Refactor Ollama service implementation: replace old service classes with new versions, update dependencies, and enhance API endpoint functionality for v1 and v2.
