+++ src/api/v2/routers/chat.py
@@ -3,12 +3,9 @@ from typing import Union
 from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import StreamingResponse
 
+from src.api.v2.ollama_service_v2 import OllamaServiceV2
 from src.api.v2.schemas.request import ChatRequest
 from src.api.v2.schemas.response import ChatResponse
-from src.api.v2.services.ollama_service import (
-    OllamaServiceV2,
-    get_ollama_service_v2,
-)
 
 router = APIRouter(
     prefix="/api/v2",
@@ -19,7 +16,7 @@ router = APIRouter(
 @router.post("/chat", response_model=ChatResponse)
 async def chat_completions(
     request: ChatRequest,
-    ollama_service: OllamaServiceV2 = Depends(get_ollama_service_v2),
+    ollama_service: OllamaServiceV2 = Depends(OllamaServiceV2.get_instance),
 ) -> Union[ChatResponse, StreamingResponse]:
     """
     Chat completions endpoint compatible with OpenAI API.
