+++ src/api/v2/services/__init__.py
@@ -2,6 +2,6 @@
 Services for v2 API endpoints.
 """
 
-from .ollama_service import OllamaServiceV2, get_ollama_service_v2
+from ..ollama_service_v2 import OllamaServiceV2
 
-__all__ = ["OllamaServiceV2", "get_ollama_service_v2"]
+__all__ = ["OllamaServiceV2"]
