+++ tests/db/middlewares/test_logging_middleware.py
@@ -5,7 +5,7 @@ from httpx import AsyncClient
 from sqlalchemy.orm import Session
 from starlette.responses import StreamingResponse
 
-from src.api.v1.services.ollama_service import GenerateResponse
+from src.api.v1.ollama_service_v1 import GenerateResponse
 from src.logs.models import Log
 
 # Mark all tests in this file as asyncio
